+++ tests/asfald.rs
@@ -22,13 +22,13 @@ fn snd_url(path: &str) -> String {
 #[test]
 // Test successful download without any flag
 fn file_with_valid_checksum() {
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg(url("/valid/the_file.txt"));
     // spawn will display the output of the command
     //cmd.spawn().unwrap();
     cmd.assert()
         .success()
-        .stdout(contains("Checksum file found !"))
+        .stdout(contains("Checksum file found at localhost!"))
         .stdout(contains("File\'s checksum is valid !"));
 
     let is_file_pred = is_file();
@@ -41,7 +41,7 @@ fn file_with_valid_checksum() {
 fn file_with_valid_checksum_o() {
     // Create out dedicated directory
     let dir: PathBuf = testdir!();
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg("-o");
     // Download the file to our dedicated directory
     cmd.arg(dir.join("the_local_file.txt"));
@@ -50,7 +50,7 @@ fn file_with_valid_checksum_o() {
     //cmd.spawn().unwrap();
     cmd.assert()
         .success()
-        .stdout(contains("Checksum file found !"))
+        .stdout(contains("Checksum file found at localhost!"))
         .stdout(contains("File\'s checksum is valid !"));
 
     let is_file_pred = is_file();
@@ -65,7 +65,7 @@ fn file_with_valid_checksum_o() {
 fn file_with_valid_checksum_p_url() {
     // Create out dedicated directory
     let dir: PathBuf = testdir!();
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     // Download the file to our dedicated directory
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
@@ -77,7 +77,7 @@ fn file_with_valid_checksum_p_url() {
 
     cmd.assert()
         .success()
-        .stdout(contains("Checksum file found !"))
+        .stdout(contains("Checksum file found at localhost!"))
         .stdout(contains("File\'s checksum is valid !"));
 
     let is_file_pred = is_file();
@@ -90,7 +90,7 @@ fn file_with_valid_checksum_p_url() {
 fn file_with_valid_checksum_p_fullpath() {
     // Create out dedicated directory
     let dir: PathBuf = testdir!();
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     // Download the file to our dedicated directory
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
@@ -102,7 +102,7 @@ fn file_with_valid_checksum_p_fullpath() {
 
     cmd.assert()
         .success()
-        .stdout(contains("Checksum file found !"))
+        .stdout(contains("Checksum file found at localhost!"))
         .stdout(contains("File\'s checksum is valid !"));
 
     let is_file_pred = is_file();
@@ -115,7 +115,7 @@ fn file_with_valid_checksum_p_fullpath() {
 fn file_with_valid_checksum_p_file_pattern() {
     // Create out dedicated directory
     let dir: PathBuf = testdir!();
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     // Download the file to our dedicated directory
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
@@ -127,7 +127,7 @@ fn file_with_valid_checksum_p_file_pattern() {
 
     cmd.assert()
         .success()
-        .stdout(contains("Checksum file found !"))
+        .stdout(contains("Checksum file found at localhost!"))
         .stdout(contains("File\'s checksum is valid !"));
 
     let is_file_pred = is_file();
@@ -140,7 +140,7 @@ fn file_with_valid_checksum_p_file_pattern() {
 fn file_with_valid_checksum_p_path_pattern() {
     // Create out dedicated directory
     let dir: PathBuf = testdir!();
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     // Download the file to our dedicated directory
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
@@ -154,7 +154,7 @@ fn file_with_valid_checksum_p_path_pattern() {
 
     cmd.assert()
         .success()
-        .stdout(contains("Checksum file found !"))
+        .stdout(contains("Checksum file found at localhost!"))
         .stdout(contains("File\'s checksum is valid !"));
 
     let is_file_pred = is_file();
@@ -167,7 +167,7 @@ fn file_with_valid_checksum_p_path_pattern() {
 fn file_p_pattern_with_http() {
     // Create out dedicated directory
     let dir: PathBuf = testdir!();
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     // Download the file to our dedicated directory
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
@@ -179,7 +179,7 @@ fn file_p_pattern_with_http() {
 
     cmd.assert()
         .success()
-        .stdout(contains("Checksum file found !"))
+        .stdout(contains("Checksum file found at localhost!"))
         .stdout(contains("File\'s checksum is valid !"));
 
     let is_file_pred = is_file();
@@ -192,7 +192,7 @@ fn file_p_pattern_with_http() {
 fn file_with_valid_checksum_q() {
     // Create out dedicated directory
     let dir: PathBuf = testdir!();
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
     cmd.arg("-q");
@@ -212,7 +212,7 @@ fn file_with_valid_checksum_q() {
 fn file_without_checksums_file() {
     // Create out dedicated directory
     let dir: PathBuf = testdir!();
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
     // Download the file to our dedicated directory
@@ -221,7 +221,7 @@ fn file_without_checksums_file() {
     //cmd.spawn().unwrap();
     cmd.assert()
         .failure()
-        .stdout(contains("Checksum file found !").not())
+        .stdout(contains("Checksum file found at localhost!").not())
         .stdout(contains("File\'s checksum is valid !").not())
         .stderr(contains("Unable to fetch checksum file"));
 
@@ -235,7 +235,7 @@ fn file_without_checksums_file() {
 fn file_without_checksums_file_but_force_absent() {
     // Create out dedicated directory
     let dir: PathBuf = testdir!();
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
     cmd.arg("-f");
@@ -245,7 +245,7 @@ fn file_without_checksums_file_but_force_absent() {
     //cmd.spawn().unwrap();
     cmd.assert()
         .success()
-        .stdout(contains("Checksum file found !").not())
+        .stdout(contains("Checksum file found at localhost!").not())
         .stdout(contains("File\'s checksum is valid !").not())
         .stdout(contains(
             "Checksum file not found, but continuing due to --force-absent or --force-invalid flag",
@@ -262,7 +262,7 @@ fn file_without_checksums_file_but_force_absent() {
 fn file_without_checksums_file_but_force_invalid() {
     // Create out dedicated directory
     let dir: PathBuf = testdir!();
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
     cmd.arg("-F");
@@ -272,7 +272,7 @@ fn file_without_checksums_file_but_force_invalid() {
     //cmd.spawn().unwrap();
     cmd.assert()
         .success()
-        .stdout(contains("Checksum file found !").not())
+        .stdout(contains("Checksum file found at localhost!").not())
         .stdout(contains("File\'s checksum is valid !").not())
         .stdout(contains(
             "Checksum file not found, but continuing due to --force-absent or --force-invalid flag",
@@ -287,11 +287,11 @@ fn file_without_checksums_file_but_force_invalid() {
 #[test]
 // File downloaded is present in checksums file, but the checksum is different
 fn file_with_invalid_checksum() {
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg(url("/invalid_checksum/the_file.txt"));
     cmd.assert()
         .failure()
-        .stdout(contains("Checksum file found !"))
+        .stdout(contains("Checksum file found at localhost!"))
         .stdout(contains("File\'s checksum is invalid !"));
 }
 
@@ -299,12 +299,12 @@ fn file_with_invalid_checksum() {
 // File downloaded is present in checksums file, but the checksum is different
 // With --force-absent: should validate the checksum if it is present
 fn file_with_invalid_checksum_force_absent() {
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg(url("/invalid_checksum/the_file.txt"));
     cmd.arg("-f");
     cmd.assert()
         .failure()
-        .stdout(contains("Checksum file found !"))
+        .stdout(contains("Checksum file found at localhost!"))
         .stdout(contains("File\'s checksum is invalid !"));
 }
 
@@ -312,12 +312,12 @@ fn file_with_invalid_checksum_force_absent() {
 // File downloaded is present in checksums file, but the checksum is different
 // With --force-absent: should validate the checksum if it is present
 fn file_with_invalid_checksum_force_invalid() {
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg(url("/invalid_checksum/the_file.txt"));
     cmd.arg("-F");
     cmd.assert()
         .success()
-        .stdout(contains("Checksum file found !"))
+        .stdout(contains("Checksum file found at localhost!"))
         .stdout(contains("File\'s checksum is invalid !"))
         .stdout(contains("⚠️⚠️ WARNING: this is insecure, and still downloads file with a checksum present, but invalid! ⚠️⚠️"));
 }
@@ -325,13 +325,13 @@ fn file_with_invalid_checksum_force_invalid() {
 #[test]
 // Test successful download without any flag
 fn file_with_path_and_valid_checksum() {
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg(url("/checksums_with_path/the_file.txt"));
     // spawn will display the output of the command
     //cmd.spawn().unwrap();
     cmd.assert()
         .success()
-        .stdout(contains("Checksum file found !"))
+        .stdout(contains("Checksum file found at localhost!"))
         .stdout(contains("File\'s checksum is valid !"));
 
     let is_file_pred = is_file();
@@ -342,11 +342,11 @@ fn file_with_path_and_valid_checksum() {
 #[test]
 // Test successful download without any flag
 fn file_with_binary_indicator() {
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg(url("/checksums_with_binary_indicator/the_file.txt"));
     cmd.assert()
         .success()
-        .stdout(contains("Checksum file found !"))
+        .stdout(contains("Checksum file found at localhost!"))
         .stdout(contains("File\'s checksum is valid !"));
 
     let is_file_pred = is_file();
@@ -361,7 +361,7 @@ fn cli_with_hash_flag() {
     let dir: PathBuf = testdir!();
 
     // Test with the right hash value
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
     cmd.arg("--hash");
@@ -372,7 +372,7 @@ fn cli_with_hash_flag() {
     //cmd.spawn().unwrap();
     cmd.assert()
         .success()
-        .stdout(contains("Checksum file found !").not())
+        .stdout(contains("Checksum file found at localhost!").not())
         .stdout(contains("Using hash passed as argument"))
         .stdout(contains("File\'s checksum is valid !"))
         .stderr(contains("Unable to fetch checksum file").not());
@@ -383,7 +383,7 @@ fn cli_with_hash_flag() {
     let _ = std::fs::remove_file(Path::new(&dir.join("the_file.txt")));
 
     // Test with a wrong hash value on the CLI: should be rejected
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
     cmd.arg("--hash");
@@ -394,7 +394,7 @@ fn cli_with_hash_flag() {
     //cmd.spawn().unwrap();
     cmd.assert()
         .failure()
-        .stdout(contains("Checksum file found !").not())
+        .stdout(contains("Checksum file found at localhost!").not())
         .stdout(contains("Using hash passed as argument"))
         .stdout(contains("File\'s checksum is invalid !"))
         .stderr(contains("Unable to fetch checksum file").not());
@@ -404,7 +404,7 @@ fn cli_with_hash_flag() {
 
     // Test with a checksum file on the server with the wrong value, but pass the right value on the
     // CLI
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
     cmd.arg(url("/invalid_checksum/the_file.txt"));
@@ -413,7 +413,7 @@ fn cli_with_hash_flag() {
     cmd.assert()
         .success()
         .stdout(contains("Using hash passed as argument"))
-        .stdout(contains("Checksum file found !").not())
+        .stdout(contains("Checksum file found at localhost!").not())
         .stdout(contains("File\'s checksum is valid !"));
     // Check file was downloaded
     assert!(is_file_pred.eval(Path::new(&dir.join("the_file.txt"))));
@@ -421,7 +421,7 @@ fn cli_with_hash_flag() {
     //
     // Test with a wrong hash value on the CLI while the server has a checksums file
     // with the right hash. This should be rejected.
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
     cmd.arg("--hash");
@@ -432,7 +432,7 @@ fn cli_with_hash_flag() {
     //cmd.spawn().unwrap();
     cmd.assert()
         .failure()
-        .stdout(contains("Checksum file found !").not())
+        .stdout(contains("Checksum file found at localhost!").not())
         .stdout(contains("Using hash passed as argument"))
         .stdout(contains("File\'s checksum is invalid !"))
         .stderr(contains("Unable to fetch checksum file").not());
@@ -448,7 +448,7 @@ fn cli_with_hash_and_p_flags() {
     let dir: PathBuf = testdir!();
 
     // Test with the right hash value
-    let mut cmd = Command::new("target/debug/asfd");
+    let mut cmd = Command::new("target/debug/asfald");
     cmd.arg("-o");
     cmd.arg(dir.join("the_file.txt"));
     cmd.arg("--hash");